Update bazel to match only extension or whole name
The hook was originally matching all files in the repository as long as they had `BUILD` or `BUILD.bazel` in them. This meant that they actually overrode the `files` setting in the user's repository. This commit ensures that, for example, `BUILD.bazel` and `whatever/BUILD.bazel` match, but not `whateverBUILD.bazelasdf`. It also adds support for `whatever.BUILD` files.